--- a/test/controllers/evaluations_controller_test.rb
+++ b/test/controllers/evaluations_controller_test.rb
@@ -13,10 +13,10 @@ class EvaluationsControllerTest < ActionDispatch::IntegrationTest
         create :submission, exercise: ex, user: user, course: @series.course, status: :correct, created_at: Time.current - 1.hour
       end
     end
-    @no_submission_user = create :user
+    @no_submission_user = users(:student)
     @no_submission_user.enrolled_courses << @series.course
     @users = @submitted_users + [@no_submission_user]
-    @course_admin = create(:staff)
+    @course_admin = users(:staff)
     @course_admin.administrating_courses << @series.course
     sign_in @course_admin
   end