… Gemfile.lock
… README.md
… test/controllers/judges_controller_test.rb
… test/controllers/notifications_controller_test.rb
… test/controllers/pages_controller_test.rb
… test/controllers/programming_languages_controller_test.rb
… test/controllers/repositories_controller_test.rb
… test/controllers/rights_requests_controller_test.rb
… test/controllers/score_items_controller_test.rb
… test/controllers/scores_controller_test.rb
… test/controllers/series_controller_test.rb
… test/controllers/submissions_controller_test.rb
… test/controllers/users_controller_test.rb
… test/factories/activity_read_state.rb
… test/factories/annotations.rb
… test/factories/api_tokens.rb
… test/factories/content_pages.rb
… test/factories/events.rb
… test/factories/exercises.rb
… test/factories/notifications.rb
… test/factories/repositories.rb
… test/factories/rights_requests.rb
… test/factories/score_items.rb
… test/factories/scores.rb
… test/factories/series.rb
… test/factories/submissions.rb
… test/factories/users.rb
… test/fixtures/courses.yml
… test/fixtures/exercises.yml
… test/fixtures/institutions.yml