This pull request tries to improve test performance. 

Running test-prof shows that over half of the time is spent in factories. We will try to remedy this in several ways:
- improve factories to create fewer associations
- add fixtures where possible
- go over the test in detail to see if all setup is needed
- replace create by build where possible

**Before**
![image](https://user-images.githubusercontent.com/481872/131477109-1698343a-872f-4659-9874-cdc225771b6d.png)

```

[TEST PROF INFO] Factories usage

 Total: 13895
 Total top-level: 4280
 Total time: 04:02.756 (out of 06:35.342)
 Total uniq factories: 37

   total   top-level     total time      time per call      top-level time               name

    2703          58        7.5253s            0.0028s             0.1457s        institution
    1535          62       15.0473s            0.0098s             0.6016s              judge
    1473         109       35.9524s            0.0244s            11.5190s         repository
    1464         435       13.9845s            0.0096s             3.9024s               user
    1293          11        3.9118s            0.0030s             0.0266s programming_language
    1283         262       81.8711s            0.0638s            13.9337s           exercise
     806         806       84.8227s            0.1052s            84.8227s         submission
     592         155       77.3136s            0.1306s            17.2639s correct_submission
     543         325       22.1759s            0.0408s            19.1588s             course
     496         496        4.1697s            0.0084s             4.1697s            student
     384         269       59.3347s            0.1545s            38.7392s             series
     381         381        3.2038s            0.0084s             3.2038s              staff
     280         280        2.3202s            0.0083s             2.3202s               zeus
      97          97        1.1388s            0.0117s             1.1388s           question
      90          90        1.5297s            0.0170s             1.5297s         score_item
      60          60       30.5477s            0.5091s            30.5477s         evaluation
      59          59        0.9644s            0.0163s             0.9644s              score
      48          48        4.7336s            0.0986s             4.7336s   wrong_submission
      42          15        0.7341s            0.0175s             0.3631s       content_page
      33          33        0.1048s            0.0032s             0.1048s           identity
      33          33        0.9652s            0.0292s             0.9652s activity_read_state
      32          32        0.2318s            0.0072s             0.2318s           provider
      26          26        0.1908s            0.0073s             0.1908s         annotation
      18          18        0.1475s            0.0082s             0.1475s       lti_provider
      16          16        0.1326s            0.0083s             0.1326s office365_provider
      16          16        0.1549s            0.0097s             0.1549s smartschool_provider
      15          15        0.1157s            0.0077s             0.1157s    gsuite_provider
      15          15        0.1435s            0.0096s             0.1435s          api_token
      13          13        0.0929s            0.0071s             0.0929s      saml_provider
      11          11        0.0202s            0.0018s             0.0202s              label
       9           9        0.2549s            0.0283s             0.2549s     rights_request
       7           7        0.7576s            0.1082s             0.7576s  course_submission
       6           6        0.0312s            0.0052s             0.0312s     temporary_user
       5           1        0.0585s            0.0117s             0.0087s             export
       5           5        0.2369s            0.0474s             0.2369s evaluation_exercise
       4           4        0.0715s            0.0179s             0.0715s       notification
       2           2        0.0108s            0.0054s             0.0108s    activity_status
```

**After**
![image](https://user-images.githubusercontent.com/481872/131884437-25689963-9407-4f70-a2e2-5617b6a4bed9.png)

```
[TEST PROF INFO] Factories usage

 Total: 2973
 Total top-level: 2132
 Total time: 00:41.293 (out of 02:13.412)
 Total uniq factories: 37

   total   top-level     total time      time per call      top-level time               name

     396         124        9.2689s            0.0234s             2.2574s           exercise
     389         147       10.8842s            0.0280s             3.4910s correct_submission
     355         219        2.1821s            0.0061s             1.1838s               user
     317         262       10.0952s            0.0318s             8.0331s             series
     267         267        6.7233s            0.0252s             6.7233s         submission
     151         140        2.1120s            0.0140s             2.0212s             course
     127         127        0.5323s            0.0042s             0.5323s            student
     120         120        0.5075s            0.0042s             0.5075s              staff
     115          23        0.1942s            0.0017s             0.0374s        institution
      91          90        5.3420s            0.0587s             5.3373s         repository
      72          72        0.4102s            0.0057s             0.4102s         score_item
      67          67        0.2378s            0.0035s             0.2378s           question
      61          44        0.4190s            0.0069s             0.3006s              judge
      49          49        6.6304s            0.1353s             6.6304s         evaluation
      49          49        0.2608s            0.0053s             0.2608s              score
      40          40        1.2446s            0.0311s             1.2446s   wrong_submission
      33          33        0.6736s            0.0204s             0.6736s activity_read_state
      33          33        0.0692s            0.0021s             0.0692s           identity
      27          27        0.1333s            0.0049s             0.1333s           provider
      24          13        0.2171s            0.0090s             0.0833s       content_page
      20          20        0.0889s            0.0044s             0.0889s               zeus
      19          19        0.1031s            0.0054s             0.1031s         annotation
      18          18        0.1155s            0.0064s             0.1155s       lti_provider
      16          16        0.0775s            0.0048s             0.0775s office365_provider
      16          16        0.0794s            0.0050s             0.0794s smartschool_provider
      15          15        0.0792s            0.0053s             0.0792s    gsuite_provider
      14          14        0.0472s            0.0034s             0.0472s          api_token
      13          13        0.0781s            0.0060s             0.0781s      saml_provider
      11          11        0.0194s            0.0018s             0.0194s programming_language
      11          11        0.0147s            0.0013s             0.0147s              label
       9           9        0.1590s            0.0177s             0.1590s     rights_request
       6           6        0.1636s            0.0273s             0.1636s  course_submission
       6           6        0.0201s            0.0033s             0.0201s     temporary_user
       5           1        0.0143s            0.0029s             0.0020s             export
       5           5        0.0434s            0.0087s             0.0434s evaluation_exercise
       4           4        0.0257s            0.0064s             0.0257s       notification
       2           2        0.0090s            0.0045s             0.0090s    activity_status
```