@bmesuere 2021-08-31
Most of the factories are now optimized. The next step is going over the tests.
![image](https://user-images.githubusercontent.com/481872/131540490-9b131384-8dd0-4d54-9574-020f0af347f0.png)


```
[TEST PROF INFO] Factories usage

 Total: 5800
 Total top-level: 4240
 Total time: 01:31.029 (out of 03:40.175)
 Total uniq factories: 37

   total   top-level     total time      time per call      top-level time               name

     806         806       18.0200s            0.0224s            18.0200s         submission
     688         421        5.3123s            0.0077s             2.8225s               user
     618         247       22.4584s            0.0363s             4.7895s           exercise
     592         155       22.4001s            0.0378s             4.7899s correct_submission
     543         325       11.7923s            0.0217s             9.0615s             course
     496         496        2.5912s            0.0052s             2.5912s            student
     384         269       25.1697s            0.0655s            15.8759s             series
     381         381        2.0298s            0.0053s             2.0298s              staff
     280         280        1.5473s            0.0055s             1.5473s               zeus
     157          54        0.4009s            0.0026s             0.1101s        institution
     110         109        9.0501s            0.0823s             9.0450s         repository
      97          97        0.4489s            0.0046s             0.4489s           question
      90          90        0.6898s            0.0077s             0.6898s         score_item
      73          56        0.6565s            0.0090s             0.5220s              judge
      60          60       13.9723s            0.2329s            13.9723s         evaluation
      59          59        0.4670s            0.0079s             0.4670s              score
      48          48        1.6266s            0.0339s             1.6266s   wrong_submission
      41          14        0.4463s            0.0109s             0.0895s       content_page
      33          33        0.8273s            0.0251s             0.8273s activity_read_state
      33          33        0.0834s            0.0025s             0.0834s           identity
      32          32        0.1714s            0.0054s             0.1714s           provider
      26          26        0.1662s            0.0064s             0.1662s         annotation
      18          18        0.1342s            0.0075s             0.1342s       lti_provider
      16          16        0.0881s            0.0055s             0.0881s smartschool_provider
      16          16        0.0847s            0.0053s             0.0847s office365_provider
      15          15        0.0822s            0.0055s             0.0822s    gsuite_provider
      15          15        0.1513s            0.0101s             0.1513s          api_token
      13          13        0.0923s            0.0071s             0.0923s      saml_provider
      11          11        0.0289s            0.0026s             0.0289s programming_language
      11          11        0.0212s            0.0019s             0.0212s              label
       9           9        0.1662s            0.0185s             0.1662s     rights_request
       7           7        0.2916s            0.0417s             0.2916s  course_submission
       6           6        0.0235s            0.0039s             0.0235s     temporary_user
       5           5        0.0834s            0.0167s             0.0834s evaluation_exercise
       5           1        0.0137s            0.0027s             0.0037s             export
       4           4        0.0204s            0.0051s             0.0204s       notification
       2           2        0.0109s            0.0054s             0.0109s    activity_status
```