@bmesuere 2021-09-01
After going over the model tests, this is the result:

```
[TEST PROF INFO] Factories usage

 Total: 4552
 Total top-level: 3316
 Total time: 01:03.972 (out of 02:52.600)
 Total uniq factories: 37

   total   top-level     total time      time per call      top-level time               name

     541         155       18.3567s            0.0339s             4.7396s correct_submission
     532         184       17.5621s            0.0330s             3.5094s           exercise
     524         290        3.7573s            0.0072s             1.7162s               user
     487         487        2.2346s            0.0046s             2.2346s            student
     378         378        1.7208s            0.0046s             1.7208s              staff
     369         265       18.6435s            0.0505s            12.3490s             series
     275         261        7.2931s            0.0265s             7.1509s             course
     272         272        1.4159s            0.0052s             1.4159s               zeus
     267         267        8.2523s            0.0309s             8.2523s         submission
     137          34        0.2951s            0.0022s             0.0640s        institution
      97          97        0.4250s            0.0044s             0.4250s           question
      91          90        6.8471s            0.0752s             6.8414s         repository
      72          72        0.4979s            0.0069s             0.4979s         score_item
      61          44        0.5576s            0.0091s             0.3818s              judge
      49          49        8.3128s            0.1696s             8.3128s         evaluation
      49          49        0.3305s            0.0067s             0.3305s              score
      48          48        1.6108s            0.0336s             1.6108s   wrong_submission
      38          13        0.3989s            0.0105s             0.0797s       content_page
      33          33        0.0820s            0.0025s             0.0820s           identity
      33          33        0.8317s            0.0252s             0.8317s activity_read_state
      27          27        0.1432s            0.0053s             0.1432s           provider
      20          20        0.1175s            0.0059s             0.1175s         annotation
      18          18        0.1305s            0.0073s             0.1305s       lti_provider
      16          16        0.0952s            0.0060s             0.0952s office365_provider
      16          16        0.1131s            0.0071s             0.1131s smartschool_provider
      15          15        0.0896s            0.0060s             0.0896s    gsuite_provider
      14          14        0.0660s            0.0047s             0.0660s          api_token
      13          13        0.0875s            0.0067s             0.0875s      saml_provider
      11          11        0.0243s            0.0022s             0.0243s programming_language
      11          11        0.0169s            0.0015s             0.0169s              label
       9           9        0.1618s            0.0180s             0.1618s     rights_request
       7           7        0.2572s            0.0367s             0.2572s  course_submission
       6           6        0.0228s            0.0038s             0.0228s     temporary_user
       5           1        0.0187s            0.0037s             0.0078s             export
       5           5        0.0626s            0.0125s             0.0626s evaluation_exercise
       4           4        0.0213s            0.0053s             0.0213s       notification
       2           2        0.0095s            0.0048s             0.0095s    activity_status
```